perf(scheduler): use coroutine in daily tasks
The different tasks now run in parallel using coroutines
